TRC: Remove subscription failure notification
At the moment, we have no consumer for this API and there is no planned
use-case as far as I know and hence let's remove it. Currently, the callback is
used when the TRC encounters an error. However, with timeouts in place we want
to keep trying forever until the application tells us to stop and hence there
is no need for a callback.