Add Operator Signature Validation for Pruning
This commit adds operator signature validation for pruning. More
specifically, this commit does the following:
- Adds some wrapper classes and functions, declared in
concord/src/utils/openssl_crypto_utils.hpp around the OpenSSL Crypto
library in order to facilitate asymmetric key signing and verification
with ECDSA keys through this library, as it was established in previous
discussions that we think this would be an ideal scheme in consideration
of our security requirements for the pruning operator's signature and
signature verification. Note these wrapper classes and functions have
been written with the intention of being generic enough that we can use
them for other signing and verification needs in Concord should we need
more application-level signatures and signing parties in the future.
- Adds unit tests for these wrappers over the OpenSSL Crypto Library.
- Introduces a new Concord configuration parameter, specifically
pruning_operator_public_key, for a public key associated with a
privileged operator who is authorized to issue pruning commands; this
parameter is effectively required if pruning is enabled and ignored if
pruning is not enabled.
- Adds logic to the pruning state machine to require all commands to
initiate pruning be verifiably signed by the privileged operator whose
public key is configured through pruning_operator_public_key; pruning
commands that cannot be verified with this public key will be ignored.
- Removes some existing logic related to having pruning orders signed by
a client proxy, as (at least as far as I, Alex, the author of this
commit message understand), we do not intend to have pruning commands
signed by client proxies in the future; they will just be signed by the
relevant operator's application-level key.
- Modifies the existing unit tests for the pruning state machine in
order to reconcile them with the changes to that state machine made in
this commit.