Run concord with the merkle storage mechanism
Change the storage mechanism from the existing v1DirectKeyValue to the
v2MerkleTree one. Key/values will be stored in a merkle tree and data
will no longer be duplicated as raw blocks will only be created
on-demand.

An important note is that the current v2MerkleTree machanism doesn't
support empty values. This is the reason why the summarized time is not
saved in the blockchain via the SerializeSummarizedTime() method.
Rationale is that this method serializes the time via Protobuf and
certain values will be serialized as zero bytes. Removing the serialized
time is not critical as it relates to pruning and it is not enabled yet.
A future commit will support empty values in v2MerkleTree .

The following will be implemented in future commits:
 * a configuration option to support switching between v1DirectKeyValue
   to v2MerkleTree
 * proper handling of changing genesis block ID when pruning - applies
   to replica_state_sync and pruning itself
 * pruning-aware state transfer - BC-1038 and BC-2498
 * support for empty values via the v2MerkleTree mechanism

Update concord-bft submodule to the latest version.