Configuration option for pre-executing all DAML requests
The changes presented in this MR are necessary, in order to unblock E2E testing of pre-execution while waiting DA to enable this feature in the Ledger API node. The idea is to introduce a global property that forces pre-execution on every commit request, so that we don't rely on the PRE_PROCESS flag being passed from the outside.

Note: for example, this would allow us to very easily run the existing DamlTests suite with pre-execution enabled.

With this MR we introduce the following:
* add an optional configuration property called `pre_execute_all_requests` (`false` by default)
* add support for the new property in the DAML gRPC Commit Service, so that the PRE_PROCESS bit is raised for each request if `pre_execute_all_requests` is `true`
* a separate configuration template dedicated to pre-execution tests (so that we don't affect existing suites)