TRC: Refactor hashing
We make lots of turns within the code for a simple task - hashing an update.
This puts the hashing implementation into a separate file and effectively two
functions. Both functions do the exact same. However, I didn't want to merge
those functions because converting between the data types can be super
expensive. In addition, I added a test to make sure it works :)