Upgrade gRPC to v1.28.x for Concord/TRC
We have been using v1.17.x and while testing the TRC I ran into a bug that was
fixed in v1.20.x. The CMake support hasn't been great but with v1.28.x (which
is the newest) we get a proper installation. At the same time, I had to upgrade
CMake as well as fix libz dependencies for concord-bft.