TRC: Timeout blocking Read() calls
The TRC need f+1 replicas to reply with the same hash. While asking around not
all replicas might be available due to the nature of a distributed system. The
current Read() calls are blocking. To unlbock the TRC, this change wraps the
blocking call into an asynchronous one with a timeout attached.