DAML Ledger: Wait for 2f+1 concords
This is a workaround but no throw away code. Based on the connection state of
the first provided Concord node the Ledger starts sending requests. With this
change we delay requests until we are able to see at least 2f+1 Concord nodes
(and their DAML gRPC service endpoint). Reading and writing in the DAML Ledger
API service are two separate and parallel tasks and hence waiting within the
TRC (read path) isn't enough to delay the first requests.