TRC: Don't test the facade
The ThinReplicaClientFacade is used for the DAML JNI integration. The interface
focuses on minimal dependecies, specifically to not include any protobuf/gRPC
headers or libraries. The facade forwards any call to the underlying TRC
implementation. However, its constructor sets up the gRPC connection. For
testing purposes this constructor is omitted. Therefore, we only test that
forwarding from the facade to the TRC works. IMHO, we can simplify our testing
by testing the TRC object directly instead of the facade. This will allow us to
write tests more conveniently w/o the need to adhere to the DAML specific
minimal interface.