TRC: Change log level at runtime
This change replaces a logging hack with another logging hack. The TRC is a
library and the logging infrastructure should be initialized by the main
application. However, the Ledger API service doesn't use the same framework. In
this change, we create a logging thread that periodically checks the logging
configuration. This improves live-debuggability.