Reject bad values for defaultable Concord params
This commit addresses the bug BC-386.

In particular, this commit changes the handling of invalid explicit
values given for defaultable parameters for conc_genconfig and Concord
when loading their configuration input and configuration files,
respectively.

Prior to this commit, it was the case that, if conc_genconfig or Concord
was given a configuration input (file) containing an invalid value for a
defaultable parameter that the conc(ord/_genconfig) application is
allowed to load the default of, then the application would fail to load
that value, but then proceed normally and ultimately successfully load
its configuration by loading the default value for the parameter in
question.

After this commit, the conc(ord/_genconfig) will terminate with an
exception if it finds an invalid value for a parameter it tries to load
from its input, even if there is a valid default it could load.

The motivation for this change is that it should help those concerned
with debugging and troubleshooting Concord notice problems earlier if
Concord immediately rejects configuration with invalid values explicitly
given rather than quietly replacing them with default values, which
could yield a possibly misleading and confusing situation in which
Concord is effectively running with values different than those that
appear to be given in its configuration (input) file(s).