Fix non-deterministic serialization in TimeContract
Fix non-deterministic serialization of time samples in
TimeContract::Serialize() by using an std::map instead of an
std::unordered_map . The effect of this instance of non-determinism is
that a replica might end up with a diverging state and that could lead
to state transfer failing with an assert (see BC-2821).

What happens is that when the system starts for the first time, it will
insert into the TimeContract::samples_ std::unordered_map variable in an
order determined by the configuration manager. That will lead to a
certain iteration order (O1). If a replica is restarted and time samples
have already been saved in storage with order O1, TimeContract would now
insert into the samples_ variable with order O1. However, that would
lead to a new iteration order (O2) and, consequently, to a different
serialization output for the samples.

A side note is that this issue became aparant when switching from the
basic DB adapter to the merkle one. Reason is that the merkle DB adapter
calculates the parent block hash when adding blocks from state transfer
by fetching the parent block from local storage. Rationale is that
blocks should be added in-order for the merkle tree to work properly.
However, if the state in storage has diverged, the calculated parent
block hash will be different from the one in the incoming state transfer
block. This problem was masked by the basic DB adapter as it is adding
incoming state transfer blocks directly into storage, without
calculating the parent hash from local storage.