TRC: Metric for ignored reads
This change introduces a new metric to counter the number of ignored reads. An
update can be ignored if its block id is less than the last verified block id.
Or, a hash update is ignored if the hash is too long. Or, an update can be
ignored if the server that previously reported a hash on an update fails to
deliver the proper data.