Thread-safe ConcordCommandsHandler
So far, the (DAML) commands handler was always being executed sequentially, which allowed it to maintain some internal state during request execution (request_, response_, request_context_). However, this changes with pre-execution where multiple requests may be executed in parallel. This implies a thread-safety requirement on the commands handler.

With this MR, we remove the `request_` and `response_` state from the ConcordCommandsHandler. The only place where `request_` was actually being used was during post-execution. This case was wired a bit unnaturally via the ExecuteCommit() method. I thought it should live in a separate method called PostExecute(), which does not use the `request_` class member, and is thus thread-safe.

PS: the safety of the refactoring is guaranteed by the DamlTest unit test suite, which covers the main flows inside the DamlKvbCommandsHandler.