Move conflict detection outside of PostExecute()
With this MR, we separate conflict detection from the post-execution "hook". Importantly, we no longer create a block with a specific write set in case of detected conflict.
The unit tests are also adapted. As discussed during the design phase, we also remove the conflict checks on the write set, as well as the related unit tests.