Add sequence number to the external client wrapper
The purpose of the patch is to add support for setting a sequence number
from the ExternalBftClient (python class) which is used byt the external
client pool (class ConcordClientPool) in Concord. This is achieved in
three steps:

1. ExternalBftClient (the python class) accepts sequence number for each
request and passes it to the external client wrapper (class
ExternalClient in test dir). The sequence number is transferred over the
UDP socket as fixed size parameter.

2. The external client wrapper passes the sequence number to the client
pool.

3. The client pool is modified to accept external sequence number as an
alternative to generate it internally.