UI: misc. bug fixes July 2020, Part 1
- Updated `blockchain-wizard.component.html` to display error
message when given committer node distribution does not match
the total number of nodes. Updated `en.json` to have this phrase
among i18n.

- Updated `blockchain-wizard.component.ts` to remove `required`
flag for the form validation to stop showing zone distribution input
as red when empty. Added checking logic to `distributeZones`
function; if the nodes are already well-distributed, it will not reset
the form to do an auto-distribution.

- Replaced all occurances of `flatMap` with `mergeMap` as the
former has been deprecated, and the recommendation was to use
`mergeMap`.

- Added `getDeletedZonePlaceHolder` to `blockchain.service.ts`
to have a way to get a placeholder object for deleted zones to show
the zone as `(Deleted Zone)`

- Added `this.blockchainService.notify.subscribe` caluse to
`dashboard.component.ts` because the nodes list did not update
when new blockchain was deployed and redirected to dashboard.
Newly added `refreshData` will reset the dashboard lists and
trigger fetching of the data, based on the updated blockchain id.

- Updated `setComponents` function in `dashboard.component`
to have initial values for dash stats, which is showing with very long
delay when switched from different consortium.

- Modified `node-list.component.html` to show default cloud zones
to be NOT navigatable. Other zones will be clickable to navigate to
zone editting page.

- Modified `nodes.service.ts`. If it cannot find the zone the node
was deployed it (the zone must have been deleted), it will fetch
the deleted zone placeholder object from blockchain service, and
set the placeholder object at `zone` and `zone_name`

- Modified `zone.component.ts` to have a routing detection for
cloud zone. If trying to edit cloud zone, it will simply redirect to the
zones list page. Also, if zone object with specified `zone_id` is not
found (404 response), the component will also redirect the user
to zones list.