Committer-tools: Self-contained
With this change we make the committer tools image self-contained. Meaning,
once the image is available on a committer node, you create a running
container, login, and start executing the scripts.
The secret sauce is how we start the container. We mount concord.config and the
docker socket from the host and provide the host's IP address. However, the
user doesn't need to worry about the specifics due to `docker run` will print
the command.
The image bloats to ~430MB because we introduce docker within. However, this is
an intermediate step and in a future we want to use SSH for all commands
instead.