Make TimeContract aware of State Transfer
Make sure TimeContract uses the latest time updates from storage. That
possibly includes updates from blocks fetched by State Transfer (ST).
This change ensures that the replica that fetches blocks via ST will
have its TimeContract in sync with other replicas' TimeContracts.

The issue that this MR aims to resolve is related to the 'samples_'
member variable. Before this change, it was loaded once, on the first
call to LoadLatestSamples(). That is typically the first call to
Update(). However, if ST runs after that, without the replica stopping,
that would lead to the 'samples_' map not containing the latest updates
from blocks fetched via ST. From that point onwards, the TimeContract
will behave differently on the replica that fetches blocks via ST.

The strategy for solving the issue is to make sure TimeContract will
always use fresh samples from storage and there won't be any leftover
state in memory across requests (i.e. execute() calls). This is achieved
via the TimeContractResetter object in execute() - it is resetting the
TimeContract on every new request and in the face of exceptions.

A cleaner solution might be to have the time updates returned at once
from TimeContract and in execute()'s scope. However, the inheritance
design expects that the TimeContract instance is a member and,
therefore, can be called across methods. Further refactoring might
improve on that limitation.

Use an std::optional instead of a raw pointer for the 'samples_'
variable in order to eliminate the need for memory management.

Improve exception safety in the following cases:
 * Updating a source shouldn't leave the existing values in the
   'samples_' map in an inconsistent state. This is achieved through
   making the SampleBody type exception-safe and move-only.
 * If loading the samples from storage fails, the state of the
   TimeContract should be well-defined. Therefore, call Reset() in case
   of such a failure.