Add spans to addBlock method
Reason:
The method `addBlock` is overridden in the product and its performance is measured by Opentracing.
The initial implementation did not change the base method but used a tricky workaround based on saving the parent span between the calls of the `execute` method.

After introducing Pre-execution the method `execute` may be called in the context of different threads.
Unfortunately, spans' behavior is undefined while used together with multithreading.

Solution:
The solution is to avoid saving spans and to create them on demand which involves the change in the base API.

Major changes:
* Added SpanWrapper to all `addBlock` methods
* Extracted Time Contract functionality from the `execute` method of `ConcordCommandsHandler` to separate methods.
* Fixed a minor bug related to the span tracking time contract. It was created in the same scope as the `execute` span.
* Added spans to the `WriteEmptyBlock` methods to make sure that `addBlock` receives a parent span.
* In order to simplify the `execute` method, the incoming span is mandatory. As a result, `daml_test.cpp` was updated with a
`NoopTracer`.
* Renamed `serialize` span to `serialize_time_response`.