Conflict detection for pre-execution
This MR implements conflict detection for pre-execution based on individual "fingerprint" (block ID) of each key. Going forward, this fingerprint could be something else, such as the hash of the value.

A unit test is also added, which makes sure the command handler exits in case of conflict, without creating a block.