CommitterTools: Make almost all cmds run via ssh
In a previous change, we introduced docker as a dependency and exposed the
host's docker socket to the container in order to run some queries. This change
converts the docker queries to run via SSH. Thereby, helpers are put in place
to easily run any command or script on a remote host. Commands that don't run
via SSH are queries to the Concord configuration file which needs to be mounted
at container start. In addition, the `my_ipv4.sh` script and its env variables
are replaced with `host-ip.sh`. The image size is reduced to ~94MB.
The change was tested on an SDDC deployment by running all scripts manually.