CommitterTools: Fix ssh-exec.sh
While testing end-to-end with different commands it turned out that this script
could be improved easily. Before, we pretty much only excepted one argument as
a command. With this change, you can have spaces in between. The photon bash
complained about the spaces after the fix so now we use the more powerful
`[[]]` operator to handle spaces.
"date" vs "echo hello"