Enable linearizability checks in Hermes
Enable linearizability checks for ApolloBftTests and
SkvbcPreexecutionTests suites in Hermes. Due to a limitation of the
linearizability tracker, the checks can be performed only for the first
tests in each suite. For this reason the tests are rearranged a bit.

This patch also adds additional fields to ExternalBftClient class so
that it becomes compatible with the linearizability tracker.