Add new config param create_tee_genesis_block for Concord
The current implementation of the linearizability tracker in concord-bft
(SkvbcTracker) works only on empty blockchain, created by the tester
replica implementation in the concord-bft project.

To overcome this limitation the patch introduces new configuration
parameter, which skips the creation of genesis block when TEE is
selected as execution engine. The parameter doesn't affect the rest of
the setups (DAML, Ethereum, HL, etc).