Operator: Connect to replicas via new client & TLS
The operator is primarily a mechanism to send specific admin commands
over a TLS connection via a BFT client. These commands require
supporting [replica specific
information](https://confluence.eng.vmware.com/display/BLOC/Replica+Specific+Information+for+BFT+Clients)
and therefore must use the [new concord-bft
client](https://github.com/vmware/concord-bft/tree/master/bftclient).

This commit enhances the operator_server build target to use new client
and establish TLS connections. It does not yet send any commands. As the
operator/new client is running on the participant node we need to
provide a configuration for a port number. We also provide docker
compose related files so that we can launch a cluster of replicas and
test the operator. So far we can view the logs and see that the operator
is indeed reading the configuration properly and connecting to the replicas.

We also added new certificates for use with the new operator clients.

The config code was made into its own library, rather than being built
as part of concord, so that it could be independently linked with the
operator without needing to build all of concord.

We also removed the placeholder header files for the new client from
concord, and the real header files were reused from concord-bft. This
was necessary to do in this commit to prevent naming/import conflicts
due to the way our cmake project is structured. However, we always
intended to do this when we could. This also required some minor changes
to the external pool client.

We bumped the concord-bft version to pull in a fix to allow proper
target usage of the new bft client and an config initialization default
for the client.

Hermes: Update install instructions and fix bug

Install instructions now mention that maven must be installed.
Additionally, a longstanding bug that prevented running tests with repo names without
slashes: i.e. 'concord-core' has been patched.