Unit test ThinReplicaClient implementation.
This commit is specifically intended to unit test that the
ThinReplicaClient implementation in the ThinReplicaClient library
correctly implements its interface at a basic level from the perspective
of code consuming the ThinReplicaClient object. This commit specifically
does not include unit testing of the ThinReplicaClient's handling of
cases involving Byzantine-faulty Thin Replica Servers.