TRC: Fix example application
When running `docker run trc-example` we cannot use the default command because
the application complains that we don't provide enough servers. However, we do
provide 4 servers. The issue is that we didn't account for skipping the timeout
value. Therefore, a quick fix is to provide the default timeout in the default
docker run command.