BC-3666: Update Dependencies for Helen
Update our use of ehcache to a more recent version.  This required using the JCache model for caching.

Moved Cache configuration to a new file CacheConfig.

I have verified that the caching is working correctly in the system by tracing, but I have been unable to get a unit test to work.  For some reason the @Cacheable annotation doesn't seem to cause the CacheInterceptor to be inserted.

Some work still should be done

* In tracing the cache calls, I can see that only one of the three caches we allocate gets used.  I suspect that changes made over the course of time have rendered these useless
* I really want a unit test, but I've delayed merging this for days trying to figure one out