BC-3969: Output info from provisioning service to an external file
Currently, all information from the provisioning service (ip addresses,
etc. etc.) are written to the log file, which is not easily accessible
in dockerized environments.

There is a requirement to store this information in human-readable
form to a user-provided location, that can then be processed further
if needed.

Add support for this requirement.