BC-3139 Add integration test for Castor
Because of some issue with how Jenkins runs, there's a problem with
committing the groovy changes and the hermes tests at the same time.

The new approach suggested (see MR 2517) is to split up the changes
into 3 separate changes:

MR1, Added it to groovy disabled
MR2, Add the test python code (won't run.)
MR3, Enable the test to see it work on pipeline.

MR1 was merged in
https://gitlab.eng.vmware.com/blockchain/vmwathena_blockchain/merge_requests/2522

These changes are for MR2: The test code is added.

MR3 will follow.

Add a deployment test for Castor that:
- launches the docker-compose-castor.yml
- tests for the output file to show SUCCESS (or FAILURE)