BC-3139: Prepare hermes groovy and main.py files for Castor test
Joe (@kkeunwoo) said that we might need to split up the Castor
integration test MRs into 2 parts: one that has only the groovy
and hermes/main.py changes, and the second (after the first gets
submitted) that contains the castor test and
castor/tests_and_artifacts.json changes.

Apparently there is an issue with the pipeline refactoring done in
May 2020. I don't understand the issue.

This MR is the first of the two: the prereq for the castor integration
tests to be put in place.