Fix possible deadlock in Thin Replica Client
This commit is primarily intended to fix a possible deadlock in the
implementation of the ThinReplicaClient object, in which the
ThinReplicaClient could get stuck if a call to Read on a gRPC stream
never returned. Though the ThinReplicaClient uses an asynchronous thread
for the reads that it times out in its main thread, the destructor for
the asynchronous thread object tries to join the thread it created,
which will deadlock if that thread is stuck is in a Read call that never
returns. This commmit fixes the issue primarily by adding more calls to
grpc::ClientContext::TryCancel to ensure the ThinReplicaClient object
cancels its gRPC calls when it times them out so that the Read calls
will return unsuccessfully rather than continue blocking after the
timeout.