BC-3139: Disable CastorDeploymentTests in spec
Because of some issue with how Jenkins runs, there's a problem
with committing the groovy changes and the hermes tests at the same
time.

The new approach suggested (see MR 2517) is to split up the changes
into 3 separate changes:

    MR1, Added it to groovy disabled

    MR2, Add the test python code (won't run.)

    MR3, Enable the test to see it work on pipeline.

These changes are for MR1: The test is disabled.

MR2 and MR3 will follow.