Fix conflict detection for pre-execution
When checking a given key for conflict, we need to get its current version (`current_block_height`) from storage, and compare it to the version which was read at pre-execution time (`read_block_height`).

On the other hand, the `storage_.get()` method's first parameter (`toBlock`) defines the upper bound (in terms of version) when reading a key. Currently, `toBlock` is incorrectly set to `read_block_height`, which means that we would never discover a conflict, even if there were one.

To fix this, we need to set `toBlock` to be equal to `storage_.getLastBlock()`, to be sure we always get the latest version of a given key.

PS: additionally, to make sure a key is available in storage, we also need to make sure the returned block height is non-zero