Move all constants to a Config class.
Provide a way to override these default values via a config properties file.

A new config file 'api_test_config.properties' was introduced to override default properties.

Sense the config file changes via a Watcherservice, and re-fill the config properties.

Run the test cases in the purview of a Test Suite. Introduce this TestSuite class that sets up the stage for tests.