BC-4207: Ensure agent talks only to on-prem shell Config Service
Agent gets its configuration by talking to the config service. Castor
launches its own config service (via docker-compose-castor.yml). The
agent needs to be able to reach this config service via the ip
address of the VM running the config-service container.

Add a parameter to the docker-compose file, CONFIG_SERVICE_IP, that
needs to be set by users when they launch the script.