Fix TimeContract deviation during pre-execution
For now, time requests are always executed sequentially (no pre-execution). In case DAML pre-execution is enabled, it runs in parallel to time request processing. In the beginning of each request execution, we reset TimeContract's internal cache (`samples_`). As pre-execution runs in parallel, this sometimes may interfere with time requests being executed at the same moment, and result in non-determinism of time updates.

For this reason, the suggested approach is to not reset TimeContract's internal cache when pre-executing requests, as they won't be modifying it (only read from it).

The implication is that, given TimeContract's current "stateful" implementation, we cannot safely pre-execute TimeRequests.

This MR fixes https://digitalasset.atlassian.net/browse/KVL-395.