ReconfigurationSM: Application specific
Reconfiguration depends on the use-case, e.g., we might have different policies
for different customers as well as different implementations for open source
and closed source. With this change we mandate that the reconfiguration state
machine has to be instantiated outside of concord's commands handler. When a
reconfiguration request is handled, the generic commands handler will call the
known reconfiguration state machine based on the IReconfiguration interface.