ReconfigurationSM: Open source vs closed source
The reconfiguration library can be replaced with a completely new
implementation based on use-cases. In order to not start from scratch, there
will be an open-source implementation with basic functionality (e.g. wedge
command). Our future closed-source implementation (talking to Agent) will
extend the open-source version. Depending on the application and use-case, we
can provide a per-customer reconfiguration engine.