Read-Only replica support
Adds support for read-only replica based on the implementation in
concord-bft.

What is RO replica?
Replica which only syncs blocks and doesn't take part of the consensus.
It backups the blockchain to external S3 object store.

How to start RO replica?
Set the configuration parameter num_ro_replicas to the desired value. It
will trigger the configuration generation for read-only nodes. The
principal_ids of the RO replica are consecutive and start right after
the committer nodes.

Some implementation details worth mentioning
* New CMake option - USE_S3_OBJECT_STORE. It can be used to
enable/disable S3 support.
* detectLocalNode() function now returns pair with node id and boolean
indicating if the replica is RO or not.
* RO replica uses new service function in main.cpp. It is called
run_ro_service(). The motivation is to separate the logic of the two
replicas avoiding too many if-else logics scattered in the code. This
approach makes the patch far less invasive.