ReconfigurationSM: Replace plugins with handler functions
This change removes the plugin architecture and replaces plugins with "handle"
functions. The "plugin" aspect of the original idea is moved into a library.
Meaning, you build your own library for your use-case. If you want to plug in
your own handler you can simply inherit from the current version and override
existing handlers. If you want to introduce a handler for a new message then
you need to modify the protobuf definition. At this point, you can choose to
alternate the exisitng dispatcher or write your own.
Furthermore, ideally, each handle function as well as the reconfiguration sm
itself only takes messages that are related to its own domain. This change does
not acomplish that and we might defer this to when we introduce CMF.
In addition, the test was modified to interact via the
IReconfigurationDispatcher interface, similiar to what the concord commands
handler does.