BC-4416: Castor appliance form factor
Castor needs some runtime-artifacts (docker-compose file, etc) to
run an on-prem deployment. Save these files into the castor docker
container itself, so that they can be extracted when the appliance
is built.

Also create a directory, orchestrator-runtime, that will store these
files. The docker-compose yml files in this directory are tailored
for on-prem use, not for test like the ones under /docker.