Use objects for configuration parameter validation
This commit changes the interface to the ConcordConfiguration so that it
accepts parameter validation logic via objects instantiating extensions
of a ParameterValidator interface type rather than function pointers;
ConcordConfiguration implementation and testing code is also revised to
comply with the new interface.

This commit is part of a broader refactoring to make ConcordConfiguration's
interface more object-oriented and reduce its usage of raw pointers
(including function pointers and void pointers).