BC-4644: NPE during GET /zones with a zoneId
We are seeing NPE during the HelenDeployDamlToSDDC test.  Root cause is that there is an on prem zone with no ID created.  This can happen when the system admin creates an on prem zone.
The question is what to do in this case.  For the moment, treat is as a mismatched org id (that is, throw not found).  If we decide that these should be visible, we need to handle this in getAllAuthorized as well.

Fixed the if statement in ZoneService, and added a test in unit tests.

Forgot to commit