BC-4416: Set up a systemd service for Castor's prereq containers
Castor requires the provisioning+config service containers to be
up and running (docker-compose-castor-prereqs.yml) before a blockchain
can be deployed using the castor container (docker-compose-castor.yml).

The config service requires the IP address of the system it is running
on. This information is sent to agents running on the blockchain VMs so
they can call back to the config service.

This means that the IP address cannot be predetermined - it can only
be determined after the Castor appliance has been created. Hence
the need for a systemd service that will, at startup, launch a shell
script (orchestrator-prereqs-launcher.sh) which figures out the IP
address of the system it is running on, and launches the prepreq
services.

Create such a service.