TRS: Use single-consumer-single-producer queue
We don't have performance numbers yet but using a bounded queue instead of a
circular buffer makes sense for our use case anyways. If the circular buffer
gets overwritten when its full or if the queue is full then the TRS won't be
able to deliver consecutive updates because the commands handler had to drop
some. In that case, the TRC might complain, and if not the TRC then the Ledger
API server. In either scenario, the TRC needs to resubscribe. With this change
we cancel the stream from the TRS side if the commands handler needs to drop
any updates.