BC-4416: Firstboot change password for root and blockchain users
Run a Oneshot service that runs only on first boot of an appliance,
which sets up passwords for root and blockchain users, based on
values provided by users in guest customization.

Also fix a problem with the base image where the docker daemon does not
start, because the bridge network is not configured. Add steps to
configure the bridge network.