Add TLS support for TRS-TRC
This adds mutual TLS support for TRS-TRC connection. To summarize -
1. This commit provides a trs_tls_enable switch and thin_replica_tls_cert_path
variable that must be set in the docker config files to enable TLS for TRS.
Similarly, a trc_tls_enable and thin_replica_tls_cert_path keys have been
added to the THIN_REPLICA_SETTINGS environment variable, which are set in the
DockerfileLedgerApi. Currently, it is the responsibility of the user to set
the switch and certificate paths for the two endpoints to the same values,
else the communication would fail. By default TLS is enabled for both TRS
and TRC.
2. TRS and TRC consume self signed certificates that have been added as a part
of this commit to docker/trs_trc_tls_certs folder.