BC-4712: Add reconfiguration support to Castor
Castor now takes a deployment type as parameter:
  - PROVISION: regular provisioning of blockchain nodes
    This is the default if no type is specified.
  - RECONFIGURE: request reconfiguriation of the blockchain nodes
Users provide this option via the ORCHESTRATOR_DEPLOYMENT_TYPE
environment variable when they run docker-compose on the
docker-compose-castor.yml.

The deployment payload is specified in either the
ProvisionDescriptorDescriptorModel or ReconfigurationDescriptorModel.

The infrastructure payload is the same for both deployment types.

Currently, both models have exactly the same data, but they may
diverge in the future. The only difference now is that providedIp
fields in clients and committers are mandatory in the reconfiguration
descriptor model. This is taken care of in the validation.

To facilitate different types of deployments (provision and
reconfigure, and in the near future, deprovision) the code has been
refactored a bit, which is why the changes affect a lot of files. But
the base functionality for provisioning has not changed.